Chore #26: edit jogboId int to long

+++ domain/my/src/main/java/com/hankki/domain/my/entity/MyJogboEntity.kt
@@ -1,7 +1,7 @@
 package com.hankki.domain.my.entity
 
 data class MyJogboEntity(
-    val jogboId : Int,
+    val jogboId : Long,
     val jogboName: String ,
     var isSelected: Boolean = false
 )
