+++ feature/my/src/main/java/com/hankki/feature/my/myjogbo/MyJogboModel.kt
@@ -3,7 +3,7 @@ package com.hankki.feature.my.myjogbo
 import com.hankki.domain.my.entity.MyJogboEntity
 
 data class MyJogboModel (
-    val jogboId : Int,
+    val jogboId : Long,
     val jogboName : String,
     var jogboSelected : Boolean = false
 )
